Implemented a config manager + storage system The storage system is just a means to simplify handling of OS-dependent paths, as well as provide better safety and should assure extensibility.
The config system makes the user interactivly configure the server if the config.json is not found.
Code-wise, it  supports modular config definition, that means modules can supply
their own keys to be added into their own config 'namespace'.
For this to work, the concerned module class must define a static method 'getConfigTemplate()'
which returns a json object of the following pattern:

```js
{
    section: 'myModuleConfig'  // the "namespace"
    description: "Short description of this module's config"
    elements: [
        {
            option: 'firstKey',
            description: "This is one single config key which can have a value"
            standard: "default-value"  // this one is optional; if omitted, the user is forced to enter something
        }, // there can be an arbitrary amount of these opjects in here
        {
            //...
        }
    }
}
```